Add resources to support being skipped on destroy

+++ cloudamqp/resource_cloudamqp_plugin.go
@@ -106,6 +106,10 @@ func resourcePluginUpdate(d *schema.ResourceData, meta interface{}) error {
 }
 
 func resourcePluginDelete(d *schema.ResourceData, meta interface{}) error {
+	if skipOnDestroy == true {
+		return nil
+	}
+
 	api := meta.(*api.API)
 	return api.DeletePlugin(d.Get("instance_id").(int), d.Get("name").(string))
 }
